UserController set up, signup route established
