+++ src/server/app.js
@@ -1,6 +1,9 @@
 // Global dependencies
 const express = require('express');
 
+// Controlla's
+const UserController = require('./controllers/Users');
+
 // Configure Express application server
 const app = express();
 
@@ -10,5 +13,12 @@ app.use((err, req, res, next) => {
   res.status(500);
   res.send(err);
 });
+
+// authorize users still need to be accomplished
+app.post('/signUp', 
+  UserController.createUserMiddleWare, 
+  UserController.authorizeUserMiddleWare, 
+  UserController.startSession
+);
 
 module.exports = app
